+++ src/components/ServiceDetails/AboutService/ServiceDemo/Purchase/PurchaseDialog/DialogTitle.js
@@ -1,0 +1,24 @@
+import React from "react";
+import { withStyles } from "@material-ui/styles";
+import MuiDialogTitle from "@material-ui/core/DialogTitle";
+import IconButton from "@material-ui/core/IconButton";
+import CloseIcon from "@material-ui/icons/Close";
+import Typography from "@material-ui/core/Typography";
+
+import { useStyles } from "./styles";
+
+const DialogTitle = withStyles(useStyles)(props => {
+  const { children, classes, onClose } = props;
+  return (
+    <MuiDialogTitle disableTypography className={classes.root}>
+      <Typography variant="h6">{children}</Typography>
+      {onClose ? (
+        <IconButton aria-label="close" className={classes.closeButton} onClick={onClose}>
+          <CloseIcon />
+        </IconButton>
+      ) : null}
+    </MuiDialogTitle>
+  );
+});
+
+export default DialogTitle;
